Main画面のときだけ常時位置を取る
