+++ src/hooks/useAskLocationPermission.ts
@@ -1,7 +1,5 @@
-import * as Location from 'expo-location';
 import * as Permissions from 'expo-permissions';
 import { useState, useEffect } from 'react';
-import { LOCATION_TASK_NAME } from '../constants';
 
 const useWatchLocation = (): [Error] => {
   const [error, setError] = useState<Error>();
@@ -13,10 +11,6 @@ const useWatchLocation = (): [Error] => {
         if (status !== 'granted') {
           throw new Error(status);
         }
-        await Location.startLocationUpdatesAsync(LOCATION_TASK_NAME, {
-          accuracy: Location.Accuracy.BestForNavigation,
-          activityType: Location.ActivityType.AutomotiveNavigation,
-        });
       } catch (e) {
         setError(e);
       }
