+++ src/screens/Main/index.tsx
@@ -1,5 +1,5 @@
 import i18n from 'i18n-js';
-import React, { Dispatch, useEffect, useState, useCallback, memo } from 'react';
+import React, { useEffect, useState, useCallback, memo } from 'react';
 import {
   ActionSheetIOS,
   Dimensions,
@@ -16,8 +16,11 @@ import {
 import { useDispatch, useSelector } from 'react-redux';
 import { useKeepAwake } from 'expo-keep-awake';
 import * as Haptics from 'expo-haptics';
-
+import * as Location from 'expo-location';
+import * as TaskManager from 'expo-task-manager';
+import * as BackgroundFetch from 'expo-background-fetch';
 import { useNavigation } from '@react-navigation/native';
+import { LocationData } from 'expo-location';
 import { TrainLCDAppState } from '../../store';
 import {
   getCurrentStationLinesWithoutCurrentLine,
@@ -29,7 +32,6 @@ import useUpdateBottomState from '../../hooks/useUpdateBottomState';
 import useRefreshStation from '../../hooks/useRefreshStation';
 import useRefreshLeftStations from '../../hooks/useRefreshLeftStations';
 import useWatchApproaching from '../../hooks/useWatchApproaching';
-import { NavigationActionTypes } from '../../store/types/navigation';
 import LineBoard from '../../components/LineBoard';
 import {
   updateBottomState,
@@ -39,14 +41,33 @@ import {
   updateSelectedDirection,
   updateSelectedBound,
 } from '../../store/actions/station';
-import { StationActionTypes } from '../../store/types/station';
 import Transfers from '../../components/Transfers';
+import { LOCATION_TASK_NAME } from '../../constants';
+import { updateLocationSuccess } from '../../store/actions/location';
 
+// eslint-disable-next-line @typescript-eslint/no-unused-vars
+let globalSetBGLocation = (location: LocationData): void => undefined;
+
+const isLocationTaskDefined = TaskManager.isTaskDefined(LOCATION_TASK_NAME);
+if (!isLocationTaskDefined) {
+  TaskManager.defineTask(
+    LOCATION_TASK_NAME,
+    ({ data, error }): BackgroundFetch.Result => {
+      if (error) {
+        return BackgroundFetch.Result.Failed;
+      }
+      const { locations } = data as { locations: LocationData[] };
+      if (locations[0]) {
+        globalSetBGLocation(locations[0]);
+        return BackgroundFetch.Result.NewData;
+      }
+      return BackgroundFetch.Result.NoData;
+    }
+  );
+}
 const MainScreen: React.FC = () => {
   const navigation = useNavigation();
-  const dispatch = useDispatch<
-    Dispatch<NavigationActionTypes | StationActionTypes>
-  >();
+  const dispatch = useDispatch();
   const { selectedLine } = useSelector((state: TrainLCDAppState) => state.line);
   const { selectedDirection, arrived } = useSelector(
     (state: TrainLCDAppState) => state.station
@@ -54,6 +75,25 @@ const MainScreen: React.FC = () => {
   const { leftStations, bottomState } = useSelector(
     (state: TrainLCDAppState) => state.navigation
   );
+  const [bgLocation, setBGLocation] = useState<LocationData>();
+  globalSetBGLocation = setBGLocation;
+
+  useEffect(() => {
+    Location.startLocationUpdatesAsync(LOCATION_TASK_NAME, {
+      accuracy: Location.Accuracy.BestForNavigation,
+      activityType: Location.ActivityType.AutomotiveNavigation,
+    });
+
+    return (): void => {
+      Location.stopLocationUpdatesAsync(LOCATION_TASK_NAME);
+    };
+  }, []);
+
+  useEffect(() => {
+    if (bgLocation) {
+      dispatch(updateLocationSuccess(bgLocation));
+    }
+  }, [bgLocation, dispatch]);
 
   const handleBackButtonPress = useCallback(() => {
     dispatch(
